moved rename_workspace tests to workspace_provider_test #362
